+++ lib/constants.js
@@ -42,8 +42,8 @@ module.exports = {
 		'forReviewEnglishInterview',
 		'englishInterviewFail',
 		'englishInterviewWaitlisted',
-		'EnglisgTestCompleted',
-		'sentEnglishTestLink',
+		'EnglishTestCompleted',
+		'pendingEnglishTest',
 
 		// culture fit interviews
 		'pendingCultureFitInterview',
