+++ lib/helpers/smsTemplates.js
@@ -123,10 +123,7 @@ smsTemplates.pendingCultureFitInterview = `Abhi tak aapne NavGurukul aapne NavGu
 
 Congrats :D
 
-Bas humein aapke saath ab ek final call karni hai aapke baare mein kuch aur cheezein samajhne ke liye. Humare kisi team member se aapko jald hi yeh call jayega.
-
-`
-
+Bas humein aapke saath ab ek final call karni hai aapke baare mein kuch aur cheezein samajhne ke liye. Humare kisi team member se aapko jald hi yeh call jayega.`
 
 smsTemplates.cultureFitInterviewWaitlisted = `NavGurukul ka test India mein hazaaron students dete hain. Hum dil se koshish karte hain ki hum saare students ko software engineering padha sakein aur achi jobs dilva sakin.
