+++ lib/helpers/syncGSheet.js
@@ -204,20 +204,19 @@ module.exports = class ChanakyaGSheetSync {
     async _getHeaders() {
         // TODO: If there is a gap of a column among two headers then return an error
         return this.doc.getCells(this.mainWorkSheetIndex, {'min-row': 1, 'max-row': 1})
-            .then((cells) => {
-                _.each(cells, (cell) => {
-                    this.headers[cell._value] = {
-                        row: cell.row,
-                        col: cell.col
-                    }
-                });   
+        .then((cells) => {
+            _.each(cells, (cell) => {
+                this.headers[cell._value] = {
+                    row: cell.row,
+                    col: cell.col
+                }
             });
+        });
     }
 
     async addRows(rows) {
         // get all rows from google sheets
         let allRows = await this.getAllRows();
-        
         let sheetRows = _.map(rows, (row) => {
             let sheetRow = Array(_.keys(this.headers).length).fill(null);
             _.each(_.keys(row), (key) => {
