+++ lib/models/englishEnrolmentKey.js
@@ -15,7 +15,7 @@ module.exports = class EnglishEnrolmentKey extends Model {
 
         return Joi.object({
             id: Joi.number().integer().greater(0),
-            key: Joi.string().length(6).required(),
+            keyId: Joi.number().integer().greater(0).required(),
             startTime: Joi.date(),
             endTime: Joi.date(),
             totalMarks: Joi.number().integer(),
