+++ lib/models/englishOption.js
@@ -17,7 +17,7 @@ module.exports = class EnglishOption extends Model {
         return Joi.object({
             id: Joi.number().integer().greater(0),
             text: Joi.string().required(),
-            questionId: Joi.number().integer().greater(0),
+            questionId: Joi.number().integer().greater(0).required(),
             correct: Joi.boolean().required(),
             createdAt: Joi.date()
         });
