+++ lib/models/englishQuestions.js
@@ -17,7 +17,7 @@ module.exports = class EnglishQuestion extends Model {
         return Joi.object({
             id: Joi.number().integer().greater(0),
             question: Joi.string().required(),
-            passageId: Joi.number().integer().greater(0),
+            passageId: Joi.number().integer().greater(0).required(),
             type: Joi.number().valid( ..._.values(CONSTANTS.questions.types) ).required(), // 1 - mcq, 2 - integer
             createdAt: Joi.date()
         });
