+++ lib/models/englishTestAttempt.js
@@ -15,7 +15,7 @@ module.exports = class EnglishQuestionAttempt extends Model {
             id: Joi.number().integer().greater(0),
             enrolmentKeyId: Joi.number().integer().greater(0).required(),
             questionId: Joi.number().integer().greater(0).required(),
-            selectedOptionId: Joi.number().integer().greater(0).allow(null),
+            selectedOptionId: Joi.number().integer().greater(0).allow(null).required(),
             createdAt: Joi.date()
         });
     }
