+++ lib/services/assessment.js
@@ -331,20 +331,20 @@ module.exports = class AssessmentService extends Schmervice.Service {
         }
     }
 
-    async sendEnglishTestLink(mobile, stage, key, txn=null) {
+    async sendEnglishTestLink(mobile, stage, keyId, key, txn=null) {
         const { EnglishEnrolmentKey } = this.server.models();
         const { exotelService } = this.server.services();
-        //  recorde the students enrolment key into englishEnrolment_key table for storing students emglish test marks.
+        // recorde the students enrolment keyId into englishEnrolment_key table for storing students english test marks.
         // and for verifying validate enrolnment key. 
         await EnglishEnrolmentKey.query(txn).insert({
-            key: key
+            keyId: keyId
         })
          
         let sendEnglishTestLink = await exotelService.sendSMS(mobile, stage, key);
         //  change student stage onces send the english test link.
         let enrolmentkey = await this.validateEnrolmentKey(key)
         await this.patchStudentDetails(enrolmentkey, {
-            stage: "sentEnglishLink"
+            stage: "pendingEnglishTest"
         })
     }
 
@@ -381,7 +381,7 @@ module.exports = class AssessmentService extends Schmervice.Service {
         const { Student } = this.server.models();
         // send English test link to student when stage is pendingEnglishInterview
         if (!details.length && key.student.stage == "pendingEnglishInterview" && key.student.contacts) {
-            await this.sendEnglishTestLink(key.student.contacts[0].mobile, key.student.stage, key.key, txn )
+            await this.sendEnglishTestLink(key.student.contacts[0].mobile, key.student.stage, key.id, key.key, txn)
         }
         
         if (details.stage) {
