+++ lib/services/englishTest.js
@@ -143,7 +143,7 @@ module.exports = class EnglishTestServices extends Schmervice.Service {
          * - testNotStarted : The student has not yet started answering the test
          */
         const { EnglishEnrolmentKey } = this.server.models();
-        let testStatus = await EnglishEnrolmentKey.query(txn).where({key: key.key})
+        let testStatus = await EnglishEnrolmentKey.query(txn).where({keyId: key.id})
         if (testStatus[0].startTime && testStatus[0].endTime) {
             return {
                 keystatus:'testAnswered',
@@ -188,7 +188,7 @@ module.exports = class EnglishTestServices extends Schmervice.Service {
                 totalMarks+=1
             }
             
-            if (answers[id]!="undefined"){
+            if (answers[id] != "undefined"){
                 QuestionAttempt.push({
                     enrolmentKeyId: key.id,
                     questionId: id,
@@ -198,19 +198,20 @@ module.exports = class EnglishTestServices extends Schmervice.Service {
                 QuestionAttempt.push({
                     enrolmentKeyId: key.id,
                     questionId: id,
+                    selectedOptionId: null
                 })
             }
         })
         await EnglishQuestionAttempt.query(txn).insertGraph(QuestionAttempt)
 
         await EnglishEnrolmentKey.query(txn).update({
-            key: key.key,
+            keyId: key.keyId,
             totalMarks: totalMarks,
             endTime: new Date()
         }).where({
             id: key.id
         })
-        
+
         return true
     }
 }
