+++ questions/english/passage1/questions.md
@@ -4,7 +4,7 @@
 {
   "type": "mcq",
   "correctOption": 3,
-  "id": 2
+  "id": 1
 }
 ```
 
@@ -15,18 +15,18 @@ About how many tigers are in the wild today?
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | 100,000 tigers | 5 |
-| 2 | 50,000 tigers | 6 |
-| 3 | 3,200 tigers | 7 |
-| 4 | 300 tigers | 8 |
-| 5 | Don't know | 54 |
+| 1 | 100,000 tigers | 2 |
+| 2 | 50,000 tigers | 1 |
+| 3 | 3,200 tigers | 6 |
+| 4 | 300 tigers | 5 |
+| 5 | Don't know | 57 |
 
 # Question Choice 2
 ```json
 {
   "type": "mcq",
   "correctOption": 3,
-  "id": 1
+  "id": 2
 }
 ```
 
@@ -37,11 +37,11 @@ The passage describes the problem of tigers being wiped out. What is one possibl
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | increasing the demand for tigers’ fur and meat | 1 |
-| 2 | allowing hunters to enter national wildlife parks | 3 |
-| 3 | making people more aware of the dangers that tigers face | 4 |
-| 4 | increasing the use of tiger bones in medicine | 2 |
-| 5 | Don't know | 53 |
+| 1 | increasing the demand for tigers’ fur and meat | 8 |
+| 2 | allowing hunters to enter national wildlife parks | 7 |
+| 3 | making people more aware of the dangers that tigers face | 9 |
+| 4 | increasing the use of tiger bones in medicine | 4 |
+| 5 | Don't know | 58 |
 
 # Question Choice 3
 ```json
@@ -59,17 +59,17 @@ One reason that tigers are going extinct is because they cannot get enough food 
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | "In Indonesia and Malaysia, [...] loggers are chopping down the jungles where tigers live to collect palm oil from plants." | 9 |
-| 2 | "In China, poachers prowl the forests and national parks on the hunt for wild tigers." | 10 |
-| 3 | "In Russia, loggers are cutting down trees to make paper. Less forest means less room for tigers to search for food." | 11 |
-| 4 | "The oil [from plants] is used in products such as ice cream, lipstick, and laundry detergent." | 12 |
-| 5 | Don't know | 59 |
+| 1 | "In Indonesia and Malaysia, [...] loggers are chopping down the jungles where tigers live to collect palm oil from plants." | 15 |
+| 2 | "In China, poachers prowl the forests and national parks on the hunt for wild tigers." | 16 |
+| 3 | "In Russia, loggers are cutting down trees to make paper. Less forest means less room for tigers to search for food." | 20 |
+| 4 | "The oil [from plants] is used in products such as ice cream, lipstick, and laundry detergent." | 19 |
+| 5 | Don't know | 61 |
 
 # Question Choice 4
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 2,
   "id": 4
 }
 ```
@@ -81,17 +81,17 @@ Based on the information in the text, why might leaders be working to reduce the
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | Leaders hope that poachers would be more likely to kill tigers if people did not want tiger fur and bones. | 29 |
-| 2 | Leaders hope that tigers poachers would be less likely to kill tigers if people did not want tiger fur and bones. | 30 |
-| 3 | Leaders hope that they can keep all tiger fur and bones if other people do not want them. | 31 |
-| 4 | Leaders hope that people will demand more tiger meat if they no longer want tiger fur and bones | 32 |
-| 5 | Don't know | 64 |
+| 1 | Leaders hope that poachers would be more likely to kill tigers if people did not want tiger fur and bones. | 3 |
+| 2 | Leaders hope that tigers poachers would be less likely to kill tigers if people did not want tiger fur and bones. | 10 |
+| 3 | Leaders hope that they can keep all tiger fur and bones if other people do not want them. | 11 |
+| 4 | Leaders hope that people will demand more tiger meat if they no longer want tiger fur and bones | 12 |
+| 5 | Don't know | 59 |
 
 # Question Choice 5
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 1,
   "id": 5
 }
 ```
@@ -103,17 +103,17 @@ What is the main idea of this passage?
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | Wild tigers are being wiped out by two major threats and could soon become extinct. | 25 |
-| 2 | Officials plan to create a system to keep tiger hunters from entering national wildlife parks. | 26 |
-| 3 | Wild tigers live in 14 countries today, but they once roamed across Asia and eastern Europe | 27 |
-| 4 | Tigers live in forests and jungles across China, India, Thailand, and eastern Russia. | 28 |
+| 1 | Wild tigers are being wiped out by two major threats and could soon become extinct. | 21 |
+| 2 | Officials plan to create a system to keep tiger hunters from entering national wildlife parks. | 22 |
+| 3 | Wild tigers live in 14 countries today, but they once roamed across Asia and eastern Europe | 23 |
+| 4 | Tigers live in forests and jungles across China, India, Thailand, and eastern Russia. | 24 |
 | 5 | Don't know | 62 |
 
 # Question Choice 6
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 2,
   "id": 6
 }
 ```
@@ -125,11 +125,11 @@ The map at the end of the text shows where tigers can be found today compared to
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | to show how large Asia and eastern Europe are | 13 |
-| 2 | to show how much the number of tigers in Asia and eastern Europe has decreased | 14 |
-| 3 | to convince readers that the number of tigers will increase in the future | 15 |
-| 4 | to convince readers that official plans to save wild tigers are working | 16 |
-| 5 | Don't know | 60 |
+| 1 | to show how large Asia and eastern Europe are | 29 |
+| 2 | to show how much the number of tigers in Asia and eastern Europe has decreased | 30 |
+| 3 | to convince readers that the number of tigers will increase in the future | 31 |
+| 4 | to convince readers that official plans to save wild tigers are working | 32 |
+| 5 | Don't know | 64 |
 
 # Question Choice 7
 ```json
@@ -147,8 +147,8 @@ Choose the answer that best completes the sentence. "Officials plan to create a 
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | instead | 17 |
-| 2 | therefore | 18 |
-| 3 | because | 19 |
-| 4 | moreover | 20 |
-| 5 | Don't know | 61 |
+| 1 | instead | 33 |
+| 2 | therefore | 34 |
+| 3 | because | 35 |
+| 4 | moreover | 36 |
+| 5 | Don't know | 65 |
