+++ questions/english/passage2/questions.md
@@ -15,11 +15,11 @@ About how many tigers are in the wild today?
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | 100,000 tigers | 21 |
-| 2 | 50,000 tigers | 22 |
-| 3 | 3,200 tigers | 23 |
-| 4 | 300 tigers | 24 |
-| 5 | Don't know | 63 |
+| 1 | 100,000 tigers | 37 |
+| 2 | 50,000 tigers | 38 |
+| 3 | 3,200 tigers | 39 |
+| 4 | 300 tigers | 40 |
+| 5 | Don't know | 66 |
 
 # Question Choice 2
 ```json
@@ -37,11 +37,11 @@ The passage describes the problem of tigers being wiped out. What is one possibl
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | increasing the demand for tigers’ fur and meat | 33 |
-| 2 | allowing hunters to enter national wildlife parks | 34 |
-| 3 | making people more aware of the dangers that tigers face | 35 |
-| 4 | increasing the use of tiger bones in medicine | 36 |
-| 5 | Don't know | 66 |
+| 1 | increasing the demand for tigers’ fur and meat | 13 |
+| 2 | allowing hunters to enter national wildlife parks | 14 |
+| 3 | making people more aware of the dangers that tigers face | 18 |
+| 4 | increasing the use of tiger bones in medicine | 17 |
+| 5 | Don't know | 60 |
 
 # Question Choice 3
 ```json
@@ -59,17 +59,17 @@ One reason that tigers are going extinct is because they cannot get enough food 
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | "In Indonesia and Malaysia, [...] loggers are chopping down the jungles where tigers live to collect palm oil from plants." | 37 |
-| 2 | "In China, poachers prowl the forests and national parks on the hunt for wild tigers." | 38 |
-| 3 | "In Russia, loggers are cutting down trees to make paper. Less forest means less room for tigers to search for food." | 39 |
-| 4 | "The oil [from plants] is used in products such as ice cream, lipstick, and laundry detergent." | 40 |
-| 5 | Don't know | 65 |
+| 1 | "In Indonesia and Malaysia, [...] loggers are chopping down the jungles where tigers live to collect palm oil from plants." | 25 |
+| 2 | "In China, poachers prowl the forests and national parks on the hunt for wild tigers." | 26 |
+| 3 | "In Russia, loggers are cutting down trees to make paper. Less forest means less room for tigers to search for food." | 27 |
+| 4 | "The oil [from plants] is used in products such as ice cream, lipstick, and laundry detergent." | 28 |
+| 5 | Don't know | 63 |
 
 # Question Choice 4
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 2,
   "id": 11
 }
 ```
@@ -81,17 +81,17 @@ Based on the information in the text, why might leaders be working to reduce the
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | Leaders hope that poachers would be more likely to kill tigers if people did not want tiger fur and bones. | 41 |
-| 2 | Leaders hope that tigers poachers would be less likely to kill tigers if people did not want tiger fur and bones. | 42 |
-| 3 | Leaders hope that they can keep all tiger fur and bones if other people do not want them. | 43 |
-| 4 | Leaders hope that people will demand more tiger meat if they no longer want tiger fur and bones | 44 |
-| 5 | Don't know | 67 |
+| 1 | Leaders hope that poachers would be more likely to kill tigers if people did not want tiger fur and bones. | 45 |
+| 2 | Leaders hope that tigers poachers would be less likely to kill tigers if people did not want tiger fur and bones. | 46 |
+| 3 | Leaders hope that they can keep all tiger fur and bones if other people do not want them. | 47 |
+| 4 | Leaders hope that people will demand more tiger meat if they no longer want tiger fur and bones | 48 |
+| 5 | Don't know | 68 |
 
 # Question Choice 5
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 1,
   "id": 12
 }
 ```
@@ -103,17 +103,17 @@ What is the main idea of this passage?
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | Wild tigers are being wiped out by two major threats and could soon become extinct. | 45 |
-| 2 | Officials plan to create a system to keep tiger hunters from entering national wildlife parks. | 46 |
-| 3 | Wild tigers live in 14 countries today, but they once roamed across Asia and eastern Europe | 47 |
-| 4 | Tigers live in forests and jungles across China, India, Thailand, and eastern Russia. | 48 |
-| 5 | Don't know | 68 |
+| 1 | Wild tigers are being wiped out by two major threats and could soon become extinct. | 41 |
+| 2 | Officials plan to create a system to keep tiger hunters from entering national wildlife parks. | 42 |
+| 3 | Wild tigers live in 14 countries today, but they once roamed across Asia and eastern Europe | 43 |
+| 4 | Tigers live in forests and jungles across China, India, Thailand, and eastern Russia. | 44 |
+| 5 | Don't know | 67 |
 
 # Question Choice 6
 ```json
 {
   "type": "mcq",
-  "correctOption": 3,
+  "correctOption": 2,
   "id": 13
 }
 ```
@@ -147,8 +147,8 @@ Choose the answer that best completes the sentence. "Officials plan to create a 
 ## Options
 | Option | Values | ID |
 |:---|:---:|:---:|
-| 1 | instead | 55 |
-| 2 | therefore | 56 |
-| 3 | because | 57 |
-| 4 | moreover | 58 |
+| 1 | instead | 53 |
+| 2 | therefore | 54 |
+| 3 | because | 55 |
+| 4 | moreover | 56 |
 | 5 | Don't know | 70 |
